Add ability to calibrate camera
Cameras often have intrisic distortions due to how 3D objects are mapped
into a 2D plane via the lens. The program created in cameraCalibration
uses the checkered pattern in /docs to calculate the calibration needed
to compensate or correct the camera. In addition to the program the
following changes are also to be made:

- Add checkered board for calibration purposes
- Modified CMakeLists to be able to compile cameraCalibration.cpp